[Health Tracking] Move settings
