+++ apps/health/ChangeLog
@@ -12,3 +12,4 @@
 0.11: Pre-minified boot&lib - folds constants and saves RAM
 0.12: Add setting for Daily Step Goal
 0.13: Add support for internationalization
+0.14: Move settings
